Change type of IntParameter getValueUnnormalised to int

+++ effect/osci_EffectParameter.h
@@ -308,7 +308,7 @@ public:
 		return getNormalisedValue(value.load());
 	}
 
-	float getValueUnnormalised() const {
+	int getValueUnnormalised() const {
 		return value.load();
 	}
